# app/Models/Category.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;




class Category extends Model
{
    protected $table = 'categories';
    public $timestamps = false;
    protected $guarded = [];

    public function transactions()
    {
        return $this->hasMany(Transaction::class);
    }
}