# docs/api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyFomin API Documentation",
        "description": "API documentation for MyFomin",
        "version": "0.0.1"
    },
    "paths": {
        "/api/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get list of all accounts",
                "operationId": "1b2191927587210525a455ee8043e241",
                "responses": {
                    "200": {
                        "description": "List of accounts returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AccountResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Create a new account",
                "operationId": "a5f891133070fbf83cdb4d0a43733c23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "err": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/accounts/{id}": {
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Update an existing account",
                "operationId": "e28ecefc685941c32052d72fdb0833cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of account to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Account"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 7000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Delete an account",
                "operationId": "38dc109b2671b3853e541b1438f1935b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of account to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Account deleted successfully"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            }
        },
        "/api/categories/{type}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get a list of categories by type (deposet or spend)",
                "description": "Returns a list of categories filtered by type. The 'type' parameter is required.",
                "operationId": "ec3297fc727a1c43c630dfc5908f7fbf",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "The type of categories to filter by",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of category by type",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing required parameter 'type'",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountRequest": {
                "title": "AccountRequest",
                "description": "Account model schema",
                "required": [
                    "name",
                    "amount"
                ],
                "properties": {
                    "name": {
                        "description": "* @OA\\Schema(\n    schema=\"AccountRequest\",\n    type=\"object\",\n    title=\"AccountRequest\",\n    description=\"Account model schema\",\n    required={\"name\", \"amount\"},",
                        "type": "string",
                        "example": "Main Account"
                    },
                    "amount": {
                        "type": "number",
                        "example": 5000
                    }
                },
                "type": "object"
            },
            "AccountResponse": {
                "title": "AccountResponse",
                "description": "Account model schema",
                "required": [
                    "id",
                    "name",
                    "amount"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Account"
                    },
                    "amount": {
                        "type": "number",
                        "example": 5000
                    }
                },
                "type": "object"
            },
            "Category": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "type": {
                        "type": "string",
                        "example": "product"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "error response interface",
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "short description"
                    },
                    "detail": {
                        "type": "string",
                        "example": "detail"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}